Refactor:StartButton분리
