+++ src/components/LandingPage/LandingDetail.jsx
@@ -1,35 +1,12 @@
 import LandingSection from './LandingSection';
-import styled from 'styled-components';
 import { Card1, Card2, CardImg } from './Card';
 import IntroSection from './IntroSection';
 import PointBox from './PointBox';
 import Title from './Title';
 import SubTitle from './SubTitle';
+import StartButton from './StartButton';
 import cardImg1 from '../../assets/image/cardImg1.png';
 import cardImg2 from '../../assets/image/cardImg2.png';
-import PrimaryButton from '../common/PrimaryButton';
-
-const StartButton = styled(PrimaryButton)`
-  padding: 14px 60px;
-  margin-bottom: 174px;
-  line-height: 2.8rem;
-  font-size: 1.8rem;
-  transition: all.3s ease;
-  span {
-    display: flex;
-    justify-content: center;
-    width: 160px;
-  }
-  &:hover {
-    background-color: var(--purple700);
-  }
-  &:active {
-    background-color: var(--purple800);
-  }
-  &:focus {
-    background-color: var(--purple800);
-  }
-`;
 
 export default function LandingDetail() {
   return (
